// src/third_party/github.com/Microsoft/TypeScript/lib/typescript.d.ts
	export interface TextRange {
		pos: number;
		end: number;
	}
	export interface ReadonlyTextRange {
		readonly pos: number;
		readonly end: number;
	}
// ...
	export interface Node extends ReadonlyTextRange {
		readonly kind: SyntaxKind;
		readonly flags: NodeFlags;
		readonly parent: Node;
	}
	export interface JSDocContainer {}
// ...
	export interface NodeArray<T extends Node> extends ReadonlyArray<T>, ReadonlyTextRange {
		readonly hasTrailingComma: boolean;
	}
	export interface Token<TKind extends SyntaxKind> extends Node {
		readonly kind: TKind;
	}
	export type EndOfFileToken = Token<SyntaxKind.EndOfFileToken> & JSDocContainer;
// ...
	export interface Declaration extends Node {
		_declarationBrand: any;
	}
// ...
	export interface Statement extends Node, JSDocContainer {
		_statementBrand: any;
	}
// ...
	export interface FileReference extends TextRange {
		fileName: string;
		resolutionMode?: SourceFile['impliedNodeFormat'];
	}
// ...
	export interface AmdDependency {
		path: string;
		name?: string;
	}
	/**
	 * Subset of properties from SourceFile that are used in multiple utility functions
	 */
	export interface SourceFileLike {
		readonly text: string;
	}
	export interface SourceFile extends Declaration {
		readonly kind: SyntaxKind.SourceFile;
		readonly statements: NodeArray<Statement>;
		readonly endOfFileToken: Token<SyntaxKind.EndOfFileToken>;
		fileName: string;
		text: string;
		amdDependencies: readonly AmdDependency[];
		moduleName?: string;
		referencedFiles: readonly FileReference[];
		typeReferenceDirectives: readonly FileReference[];
		libReferenceDirectives: readonly FileReference[];
		languageVariant: LanguageVariant;
		isDeclarationFile: boolean;
		/**
		 * lib.d.ts should have a reference comment like
		 *
		 *  /// <reference no-default-lib="true"/>
		 *
		 * If any other file has this comment, it signals not to include lib.d.ts
		 * because this containing file is intended to act as a default library.
		 */
		hasNoDefaultLib: boolean;
		languageVersion: ScriptTarget;
		/**
		 * When `module` is `Node16` or `NodeNext`, this field controls whether the
		 * source file in question is an ESNext-output-format file, or a CommonJS-output-format
		 * module. This is derived by the module resolver as it looks up the file, since
		 * it is derived from either the file extension of the module, or the containing
		 * `package.json` context, and affects both checking and emit.
		 *
		 * It is _public_ so that (pre)transformers can set this field,
		 * since it switches the builtin `node` module transform. Generally speaking, if unset,
		 * the field is treated as though it is `ModuleKind.CommonJS`.
		 *
		 * Note that this field is only set by the module resolution process when
		 * `moduleResolution` is `Node16` or `NodeNext`, which is implied by the `module` setting
		 * of `Node16` or `NodeNext`, respectively, but may be overriden (eg, by a `moduleResolution`
		 * of `node`). If so, this field will be unset and source files will be considered to be
		 * CommonJS-output-format by the node module transformer and type checker, regardless of extension or context.
		 */
		impliedNodeFormat?: ModuleKind.ESNext | ModuleKind.CommonJS;
	}
// ...
	export enum ModuleKind {
		None = 0,
		CommonJS = 1,
		AMD = 2,
		UMD = 3,
		System = 4,
		ES2015 = 5,
		ES2020 = 6,
		ES2022 = 7,
		ESNext = 99,
		Node16 = 100,
		NodeNext = 199,
	}
// ...
	export enum ScriptTarget {
		ES3 = 0,
		ES5 = 1,
		ES2015 = 2,
		ES2016 = 3,
		ES2017 = 4,
		ES2018 = 5,
		ES2019 = 6,
		ES2020 = 7,
		ES2021 = 8,
		ES2022 = 9,
		ESNext = 99,
		JSON = 100,
		Latest = 99,
	}
	export enum LanguageVariant {
		Standard = 0,
		JSX = 1,
	}
// ...
	export interface System {
		args: string[];
		newLine: string;
		useCaseSensitiveFileNames: boolean;
		write(s: string): void;
		writeOutputIsTTY?(): boolean;
		getWidthOfTerminal?(): number;
		readFile(path: string, encoding?: string): string | undefined;
		getFileSize?(path: string): number;
		writeFile(path: string, data: string, writeByteOrderMark?: boolean): void;
		/**
		 * @pollingInterval - this parameter is used in polling-based watchers and ignored in watchers that
		 * use native OS file watching
		 */
		watchFile?(
			path: string,
			callback: FileWatcherCallback,
			pollingInterval?: number,
			options?: WatchOptions
		): FileWatcher;
		watchDirectory?(
			path: string,
			callback: DirectoryWatcherCallback,
			recursive?: boolean,
			options?: WatchOptions
		): FileWatcher;
		resolvePath(path: string): string;
		fileExists(path: string): boolean;
		directoryExists(path: string): boolean;
		createDirectory(path: string): void;
		getExecutingFilePath(): string;
		getCurrentDirectory(): string;
		getDirectories(path: string): string[];
		readDirectory(
			path: string,
			extensions?: readonly string[],
			exclude?: readonly string[],
			include?: readonly string[],
			depth?: number
		): string[];
		getModifiedTime?(path: string): Date | undefined;
		setModifiedTime?(path: string, time: Date): void;
		deleteFile?(path: string): void;
		/**
		 * A good implementation is node.js' `crypto.createHash`. (https://nodejs.org/api/crypto.html#crypto_crypto_createhash_algorithm)
		 */
		createHash?(data: string): string;
		/** This must be cryptographically secure. Only implement this method using `crypto.createHash("sha256")`. */
		createSHA256Hash?(data: string): string;
		getMemoryUsage?(): number;
		exit(exitCode?: number): void;
		realpath?(path: string): string;
		setTimeout?(callback: (...args: any[]) => void, ms: number, ...args: any[]): any;
		clearTimeout?(timeoutId: any): void;
		clearScreen?(): void;
		base64decode?(input: string): string;
		base64encode?(input: string): string;
	}